% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltypecount.R
\name{celltypecount}
\alias{celltypecount}
\title{celltypecount}
\usage{
celltypecount(dataset, label)
}
\arguments{
\item{dataset}{A seurat object.}

\item{label}{Original datatype (ie. GFP+/GFP-)}
}
\value{
A dataframe of the counts in the cluster by celltype.
}
\description{
(All)Outputs the counts in the cluster by celltype given by
identify_cluster
}
\examples{
cellcounts_40pc_0.2 <- celltypecount(plaque.combined_40pc_0.2, label = "GFP")



}
