% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_averagelog2FC.R
\name{cluster_averagelog2FC}
\alias{cluster_averagelog2FC}
\title{cluster_averagelog2FC}
\usage{
cluster_averagelog2FC(dataset, weight = "na")
}
\arguments{
\item{dataset}{The data frame output of identify_cluster.R}
}
\value{
A data frame object that has avg_log2FC per cluster for each
}
\description{
Goes through given top markers and outputs the avg_log2FC per cluster for each cell type
}
\examples{
cluster2_avglog2FC <- cluster_averagelog2FC(dataset = output_table, weight = "weighted")


}
