% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_meanlogFC.R
\name{cluster_meanlogFC}
\alias{cluster_meanlogFC}
\title{cluster_meanlogFC}
\usage{
cluster_meanlogFC(dataset, weight = "na")
}
\arguments{
\item{dataset}{The output of identify_cluster(findtop_conservedmarkers())}

\item{weight}{If you want to have weights, then weight = "weighted"}
}
\value{
A dataframe with the average meanlogFC per cluster for each celltype.
}
\description{
(All) Tells you the average meanlogFC per cluster for each celltype
(weighted or not weighted )
}
\examples{
cluster_meanlogFC(clust_40pc_0.2_identifymarkers)



}
