% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctlipa_fisher.test.R
\name{correctlipa_fisher.test}
\alias{correctlipa_fisher.test}
\title{correctlipa_fisher.test}
\usage{
correctlipa_fisher.test(output, counts, alt = "two.sided")
}
\arguments{
\item{output}{A data frame from the output of lipa_expressionmarkers}

\item{counts}{A data frame from the output of originalcount (gives counts by cluster split by GFP)}

\item{alt}{For the Fisher's Test: "two.sided" or "greater" or "less" (default is two.sided)}
}
\value{

}
\description{
Test for clusters between GFP+ and GFP- whether there is signficiantly different/more/less
expression of Lipa.
}
\examples{
lipa.40pc_0.2_all <- correctlipa_fisher.test(lipa.40pc_0.2, counts_40pc_0.2)

}
