% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findall_conservedmarkers.R
\name{findall_conservedmarkers}
\alias{findall_conservedmarkers}
\title{findall_conservedmarkers}
\usage{
findall_conservedmarkers(seuratobject, clusternumber)
}
\arguments{
\item{dataset}{}
}
\value{

}
\description{

}
\examples{



}
