% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_cluster.R
\name{identify_cluster}
\alias{identify_cluster}
\title{identify_cluster}
\usage{
identify_cluster(topmarkers, upper = "na", top20 = "na", writecsv = "na")
}
\arguments{
\item{topmarkers}{A dataset of topmarkers}
}
\value{
A data frame object that has all the markers and corresponding
clusters and the cluster type for each gene.
}
\description{
Goes through given top markers and extract those that match the
cluster identifiers.
}
\examples{
output_table <- identify_cluster(topmarkers, upper = "upper", top20 = "top20", writecsv = "writecsv")


}
