% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipa_expressionmarkers.R
\name{lipa_expressionmarkers}
\alias{lipa_expressionmarkers}
\title{lipa_expressionmarkers}
\usage{
lipa_expressionmarkers(seuratobject, clusternumber, feature)
}
\arguments{
\item{seuratobject}{A seurat object.}

\item{clusternumber}{Output of top_cluster.}

\item{feature}{Name of the feature you want to find}
}
\value{
A dataframe of the avg2logFC of the feature you want expressed
in each cluster.
}
\description{
(LipaTg)Returns the expression per cluster of the feature you
want expressed sorted by avg2logFC
}
\examples{
lipa.40pc_0.2 <- lipa_expressionmarkers(plaque.combined_40pc_0.2,clust_40pc_0.2,
clusternumber = clust_40pc_0.2, feature = "Lipa")



}
