% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipaallmacro_fishertest.R
\name{lipaallmacro_fishertest}
\alias{lipaallmacro_fishertest}
\title{lipaallmacro_fishertest}
\usage{
lipaallmacro_fishertest(output, counts, alt = "two.sided")
}
\arguments{
\item{output}{Output of lipa_expressionmarkers}

\item{counts}{Output of originalcount}

\item{alt}{"two.sided", "greater", "less"}
}
\value{
A dataframe of the fisher's test output for Lipa in macrophage
subpopulations of GFP+/GFP-.
}
\description{
(LipaTg)Tests for Lipa in macrophages populations in GFP+/GFP-.
}
\examples{
lipa.40pc_0.2 <- lipaallmacro_fishertest(plaque.combined_40pc_0.2,
counts = counts_40pc_0.2, alt = "two.sided")



}
