% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/originalcount.R
\name{originalcount}
\alias{originalcount}
\title{originalcount}
\usage{
originalcount(dataset)
}
\arguments{
\item{dataset}{A seurat object.}
}
\value{
A dataframe of the counts in the cluster by original datatype
}
\description{
(All)Outputs the counts in the cluster by original datatype.
}
\examples{
counts_40pc_0.2 <- originalcount(plaque.combined_40pc_0.2)
would return counts split by GFP+/GFP-


}
