% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotcount_cluster.R
\name{plotcount_cluster}
\alias{plotcount_cluster}
\title{plotcount_cluster}
\usage{
plotcount_cluster(dataset, GFP)
}
\arguments{
\item{dataset}{A seurat object.}

\item{label}{Original datatype (ie. GFP+/GFP-)}
}
\value{
A barplot of the counts in each cluster by celltype.
}
\description{
(LipaTg)Plots the distribution of cells by cluster, split by
original datatype.
}
\examples{
plotcounts_40pc_0.2 <- plotcount_cluster(plaque.combined_40pc_0.2)



}
