% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signficiant_lipafisher.R
\name{signficiant_lipafisher}
\alias{signficiant_lipafisher}
\title{signficiant_lipafisher}
\usage{
signficiant_lipafisher(dataset)
}
\arguments{
\item{dataset}{A data frame which is the output of correctlipa_fisher.test or lipaall_fisher.test}
}
\value{
A data frame with p-value < 0.05
}
\description{
Tells you which clusters have signficiant lipa expression.
}
\examples{

signficiant_lipafisher(cluster1_correctlipafisher)

}
