% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significant_lipafisher.R
\name{significant_lipafisher}
\alias{significant_lipafisher}
\title{significant_lipafisher}
\usage{
significant_lipafisher(dataset)
}
\arguments{
\item{dataset}{The output of correctlipa_fisher.test}
}
\value{
A dataframe with the significant Lipa clusters
}
\description{
(LipaTg) Gives you which cluster have significant lipa subpopulations.
}
\examples{
lipa.40pc_0.2_sig <- significant_lipafisher(lipa.40pc_0.2_fisher)

}
