% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_cluster.R
\name{top_cluster}
\alias{top_cluster}
\title{top_cluster}
\usage{
top_cluster(data)
}
\arguments{
\item{data}{The output of cluster_averagelog2FC.}
}
\value{
A dataframe with the top clusters labeled.
}
\description{
(All) Gives you the top cluster from cluster_averagelog2FC
}
\examples{
top_cluster(dataset)



}
