% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findtop_conservedmarkers.R
\name{findtop_conservedmarkers}
\alias{findtop_conservedmarkers}
\title{findtop_conservedmarkers}
\usage{
findtop_conservedmarkers(datalist)
}
\arguments{
\item{datalist}{A data frame of conserved markers}
}
\value{
A dataframe of the top 20 conserved markers
}
\description{
(LipaTg)Finds the top 20 conserved markers to be put into
identify_cluster.
}
\examples{
clust_40pc_0.4_identifymarkers <- identify_cluster(findtop_conservedmarkers(clust_40pc_0.4_conservedmarkers))


}
