% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipa_expressionmarkers.R
\name{lipa_expressionmarkers}
\alias{lipa_expressionmarkers}
\title{lipa_expressionmarkers}
\usage{
lipa_expressionmarkers(seuratobject, clusternumber)
}
\arguments{
\item{dataset}{}
}
\value{

}
\description{

}
\examples{



}
