% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipa_fisher.test.R
\name{lipa_fisher.test}
\alias{lipa_fisher.test}
\title{lipa_fisher.test}
\usage{
lipa_fisher.test(output, counts, alt = "two.sided")
}
\arguments{
\item{dataset}{}
}
\value{

}
\description{

}
\examples{
lipa.30pc_0.4_fisher <- lipa_fisher.test(lipa.30pc_0.4, counts_30pc_0.4)


}
