% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_indents.R
\name{rename_indents}
\alias{rename_indents}
\title{rename_indents}
\usage{
rename_indents(seuratobject, labels)
}
\arguments{
\item{seuratobject}{The seurat object and the outuput of top_cluster.}

\item{labels}{The output of top_cluster}
}
\value{
N/a
}
\description{
(All)Relabels the clusters automatically for graphing.
}
\examples{
plaque.combined_40pc_0.6 <- rename_indents(plaque.combined_40pc_0.6,
top_cluster(plaque.combined_40pc_0.6clusteraverages))



}
