% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_indents.R
\name{rename_indents}
\alias{rename_indents}
\title{rename_indents}
\usage{
rename_indents(seuratobject, labels)
}
\arguments{
\item{dataset}{The seurat object and the outuput of top_cluster.}
}
\value{
N/a
}
\description{
Relabels the clusters automatically for graphing.
}
\examples{
rename_indents(seuratobject, labels)



}
